now using cogs and Cog ovjects to modularize the bot
